feat: move route unit test to router plugin
