+++ packages/runtime/plugin-runtime/tests/router/data.test.ts
@@ -2,7 +2,7 @@
  * @jest-environment node
  */
 import path from 'path';
-import { hasLoader, hasAction } from '../../src/plugins/analyze/utils';
+import { hasLoader, hasAction } from '../../src/router/cli/code/utils';
 
 describe('should verify loader and action normally', () => {
   const dataFile = path.resolve(
