+++ packages/runtime/plugin-runtime/tests/router/fixtures/nested-routes/__auth/layout.config.ts
