+++ packages/runtime/plugin-runtime/tests/router/fixtures/nested-routes/user/$.data.ts
