+++ packages/runtime/plugin-runtime/tests/router/makeLegalIdentifier.test.ts
@@ -1,4 +1,4 @@
-import { makeLegalIdentifier } from '../../src/plugins/analyze/makeLegalIdentifier';
+import { makeLegalIdentifier } from '../../src//router/cli/code/makeLegalIdentifier';
 
 describe('is default export function', () => {
   test('camel-cases names', () => {
